+++ javascript/types/game-handler.js
@@ -1,0 +1,23 @@
+/** @typedef {import('./event').EventBusPublicAPI} EventBusPublicAPI */
+/** @typedef {import('./engine').EnginePublicAPI} EnginePublicAPI */
+/** @typedef {import('./loop').LoopPublicAPI} LoopPublicAPI */
+/** @typedef {import('./audio').AudioManagerPublicAPI} AudioManagerPublicAPI */
+/** @typedef {import('./layout').LayoutManagerPublicAPI} LayoutManagerPublicAPI */
+/** @typedef {import('./settings').SettingsManagerPublicAPI} SettingsManagerPublicAPI */
+
+/**
+ * @typedef {Object} GameHandlerDependencies
+ * @property {EventBusPublicAPI} eventBus
+ * @property {EnginePublicAPI} engine
+ * @property {LoopPublicAPI} loop
+ * @property {AudioManagerPublicAPI} audioManager
+ * @property {LayoutManagerPublicAPI} layoutManager
+ * @property {SettingsManagerPublicAPI} settingsManager
+ */
+
+/**
+ * @typedef {Object} GameHandlerPublicAPI
+ * @property {() => void} registerEvents - Registers all core game events and their corresponding handlers.
+ */
+
+export {};
